+++ Source/Core/InputCommon/InputCommon.vcxproj.filters
@@ -22,6 +22,7 @@
     <ClCompile Include="GCAdapter.cpp" />
     <ClCompile Include="InputConfig.cpp" />
     <ClCompile Include="InputStabilizer.cpp" />
+    <ClCompile Include="LibusbUtils.cpp" />
     <ClCompile Include="ControllerInterface\DInput\DInput.cpp">
       <Filter>ControllerInterface\DInput</Filter>
     </ClCompile>
@@ -58,6 +59,7 @@
     <ClInclude Include="GCPadStatus.h" />
     <ClInclude Include="InputConfig.h" />
     <ClCompile Include="InputStabilizer.h" />
+    <ClCompile Include="LibusbUtils.h" />
     <ClInclude Include="ControllerInterface\DInput\DInput.h">
       <Filter>ControllerInterface\DInput</Filter>
     </ClInclude>
